requirements.txt
# Core Framework
Flask==3.0.3
Werkzeug==3.0.6

# WSGI Server
gunicorn==21.2.0

# Database
pysqlcipher3==1.2.0

# API Documentation
flasgger==0.9.7.1

# Data Processing
pandas==2.2.3
numpy==2.2.2

# Visualization
plotly==6.0.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0

# Security & Validation
cryptography==44.0.0
python-dotenv==1.0.1
marshmallow==3.23.2

# Caching & Rate Limiting
Flask-Caching==2.3.0
Flask-Limiter==3.8.0

# CORS
Flask-Cors==5.0.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
